cf. #35 - fix a memory leak
The problem is that it's not safe to repeatdly create Jinja2
environments. Otherwise said, Jinja2 environments leaks. To workaround
this bug Jinja2 environment is turned into a singleton via a global
which is ugly but I have no other smart to solve the issue except
create an app *context* when the application starts and forward that
around.